ai-attendance-system/desktop-app/requirements.txt
PyQt5==5.15.9
requests==2.31.0
Pillow==10.0.0
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2